feat(#118): removed user component from components, and moved it under routes, added api integration for onboarding page One with error validations
